create board move render channels
