+++ mysite/templates/chessMatch.html
@@ -24,7 +24,7 @@
             </button>
         </div>
     </nav>
-    <div class="d-flex p-2 justify-content-center">
+    <div class=" p-2 justify-content-center">
         <div class="chessboard">
             {% for nums in range %}
                 {% if forloop.counter|divisibleby:2 %}
@@ -58,21 +58,31 @@
         const processClickUrl = "{% url 'processClick' %}";
         const token = "{{ csrf_token }}";
         const roomName = JSON.parse(document.getElementById("room-name").textContent);
-        const playerSocket = new WebSocket("ws://" + window.location.host + "/ws/chess/match/" + roomName + "/");
+        const playerChatSocket = new WebSocket("ws://" + window.location.host + "/ws/chess/match/" + roomName + "/chat/");
+        const playerGameSocket = new WebSocket("ws://" + window.location.host + "/ws/chess/match/" + roomName + "/game/");
+        const game = new Game(new Chessboard());
 
-        playerSocket.onmessage = function (e) {
+
+        playerChatSocket.onmessage = function (e) {
             const data = JSON.parse(e.data);
             console.log(data);
             document.querySelector("#chat-log").value += data.message + "\n";
         }
-        playerSocket.onclose = function (e) {
+        playerChatSocket.onclose = function (e) {
+            console.error("Chat socket closed unexpectedly");
+        };
+        playerGameSocket.onmessage = function (e) {
+            const data = JSON.parse(e.data);
+            console.log(data);
+            game.chessboard.processBoard(data.gamedata);
+        }
+        playerGameSocket.onclose = function (e) {
             console.error("Chat socket closed unexpectedly");
         };
 
 
         
         const startGameBtn = document.getElementById("startgamebtn");
-        const game = new Game(new Chessboard());
 
         document.addEventListener("click", function(event) {
             game.setupEventListener(event);
@@ -95,7 +105,7 @@
         document.querySelector("#chat-message-submit").onclick = function (e) {
             const messageInputDom = document.querySelector("#chat-message-input");
             const message = messageInputDom.value;
-            playerSocket.send(
+            playerChatSocket.send(
             JSON.stringify({
                 message: message,
             })
